move src files into src/, other scripts into tools/
