+++ tools/setup-db.js
@@ -1,4 +1,4 @@
-require('dotenv').config();
+require('dotenv').config({ path: __dirname + '/../.env' });
 
 var MongoClient = require('mongodb').MongoClient;
